Make AutoUpdate take advantage of github api ETag
This means that if there is no change to the output
the github api is going to give us it returns NULL
with a 304 header, we can then say there are no updates.

The github api will not make this request as one of the 60
we are allowed per hour thus we can check for updates more often!