Apply code style updates to /Includes/
Ignore lime.php, since it's from an external library (?).

These are the same style updates as before, except that, in the
interests of avoiding controversy, I haven't forced the introduction
of brace if if/else statements.

Obviously the lack of tests hurts the extent to which I know this change
is 'safe', but I have looked through all the changes with a no-whitespace
clean diff to make sure nothing looks odd. ANd beside, PHPStorm is tried
and tested for safety.