Fix Image class constructor
You can't piggy-back on Image::imageinfo() to normalize the title for you,
because that information is filtered about by the list handler (since it is not
strictly imageinfo). Such filtering is arguably a very useful feature of the
imageinfo() function, so rather than edit it, I put in a second API call just
for title normalization -- should be cheap enough.